[plasma] move newElectronCrossSection and getElectronCrossSection to PlasmaElectronFactory
